Move files from calculate_statistics to a more Pythonic structure
- It is suggested to run the main.py script with `python -m` because
  stats/main.py needs to import settings from its parent directory,
  but when it is run directly as a script it doesn't know that there's
  anything above it.
